requirements.txt
fastapi==0.116.1
redis
python-keycloak==4.6.2
httpx==0.27.0
pydantic==2.10.5
pydantic-settings==2.4.0
python-multipart==0.0.6
SQLAlchemy==2.0.41
aiomysql==0.2.0
mysql-connector-python==9.4.0
uvicorn==0.35.0
gunicorn==23.0.0
